Use a pool of buffers when publishing
Before when a single buffer was shared multiple simulations calls to
AMQPChannel#basicPublish could corrupt the buffer.

Fixes #36

Downside with this fix is that if many microtasks run simulantious many
buffers will be created, but never cleaned up, so there's a risk of memory
leak. Upside compared to allocating a new buffer for each call to
basicPublish is that this way is a lot faster (about 50%).