Call socket.destroy() in closeSocket() to ignore possible RST response
The test "onerror is not called when conn is closed by client"
intermittently fails.

According to some resources, such as RFC-1122 (4.2.2.12 Closing a
Connection): if the host actively closes a connection, while still
having unread incoming data available, the host should send the signal
RST (losing any received data) instead of FIN. This may or may not be
what is happening, but there's there's something causing a RST to be
received by the client at least some of the time.

This pull request proposes the solution of calling socket.destroy() in
closeSocket() to ignore the possibility of a RST response. This was also
the solution in postwait/node-amqp#444.

Close #72